🏗️ Reorganize theme system for clarity
Split theme logic from theme definitions:
- Theme definitions stay in themes/ (data)
- Theme system code moves to lib/theme/ (logic)

New structure:
lib/theme/
├── site/         # Site theme system (provider, engine)
└── application/  # App theme system (registry, factory, css-generator)

themes/
├── site/         # Site theme definitions
└── application/  # App theme definitions only

This separation makes it clear what's code vs what's data
