+++ lib/theme/application/factory.ts
