+++ lib/theme/application/init.ts
@@ -1,9 +1,9 @@
 // Initialize built-in themes
-import { registerTheme } from './theme-registry'
-import { validateAndNormalize } from './theme-adapter'
-import { terminalTheme } from './terminal-theme'
-import { terminalThemeHybrid } from './terminal-theme-hybrid'
-import { directoryTheme } from './directory-theme'
+import { registerTheme } from './registry'
+import { validateAndNormalize } from './factory'
+import { terminalTheme } from '@/themes/application/terminal'
+import { terminalThemeHybrid } from '@/themes/application/terminal-hybrid'
+import { directoryTheme } from '@/themes/application/directory'
 // import { cyberpunkTheme } from './cyberpunk-theme'
 // import { minimalTheme } from './minimal-theme'
 // import { retroTheme } from './retro-theme'
