+++ themes/application/terminal-hybrid.ts
