+++ themes/application/terminal.ts
